fix(security): ensure --patch-env regenerates short and low-entropy secrets
ensure_env_file_secrets() only checked known-weak names and the
__REPLACE_ME__ placeholder. Settings.validate_security_combinations()
additionally hard-fails on values shorter than 32 chars or with entropy
below 3.5, leaving a gap where --patch-env would leave a non-compliant
value in .env that blocks startup anyway.

- Move calculate_entropy(), MIN_SECRET_LENGTH, and MIN_ENTROPY into
  _security_constants.py so both config.py and init_secrets.py share
  the same predicate without a circular import
- Remove the now-duplicate calculate_entropy() definition from config.py
  and drop the unused import math
- Apply the full four-class compliance predicate (empty, placeholder,
  known-weak, too-short, low-entropy) to JWT_SECRET_KEY and
  AUTH_ENCRYPTION_SECRET in ensure_env_file_secrets(); BASIC_AUTH_PASSWORD
  retains the original name/placeholder-only check since Settings never
  hard-fails on it
